bugfix for previous commit!!!!

+++ src/test/java/dev/floelly/ghostnetfishing/integration/end2end/NewNetEndToEndTest.java
@@ -62,8 +62,8 @@ class NewNetEndToEndTest extends AbstractMySQLContainerTest {
             double dbLong = rs.getDouble(DB_COLUMN_LONGITUDE);
             String dbSize = rs.getString(DB_COLUMN_SIZE);
 
-            assertEquals(Long.parseLong(randomLatitude), dbLat, 0.0001, "Latitude of first persisted net does not match input value");
-            assertEquals(Long.parseLong(randomLongitude), dbLong, 0.0001, "Longitude of first persisted net does not match input value");
+            assertEquals(Double.parseDouble(randomLatitude), dbLat, 0.0001, "Latitude of first persisted net does not match input value");
+            assertEquals(Double.parseDouble(randomLongitude), dbLong, 0.0001, "Longitude of first persisted net does not match input value");
             assertEquals(size, dbSize, "Size of first persisted net does not match input value");
             assertFalse(rs.next(), "Should not have more than one persisted net");
         }
